requirements.txt
asyncio==3.4.3
attrs==20.3.0
autopep8==2.1.0
backports.shutil-get-terminal-size==1.0.0
backports.shutil_which==3.5.2
beautifulsoup4==4.12.3
black==24.3.0
blinker==1.7.0
braceexpand==0.1.7
Brotli==1.1.0
cachelib==0.12.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
check50==3.3.11
clang-format==18.1.2
cli50==7.5.1
click==8.1.7
colorama==0.4.6
compare50==1.2.6
contourpy==1.2.1
countryinfo==0.1.2
cryptography==42.0.5
cs50==9.3.4
CS50-VSIX-Client @ file:///opt/cs50/extensions/cs50vsix-client
cssselect2==0.7.0
cycler==0.12.1
defusedxml==0.7.1
djhtml==3.0.6
EditorConfig==0.12.4
Flask==3.0.2
Flask-Session==0.8.0
fonttools==4.50.0
fpdf2==2.7.8
freezegun==1.4.0
geographiclib==2.0
geopy==2.4.1
greenlet==3.0.3
help50==3.0.5
html5lib==1.1
icdiff==2.0.7
idna==3.6
inflect==7.0.0
iniconfig==2.0.0
intervaltree==3.1.0
itsdangerous==2.1.2
jellyfish==0.11.2
Jinja2==3.1.3
jsbeautifier==1.15.1
kiwisolver==1.4.5
lib50==3.0.11
MarkupSafe==2.1.5
matplotlib==3.8.4
more-itertools==10.2.0
msgspec==0.18.6
mypy-extensions==1.0.0
natsort==8.4.0
numpy==1.26.4
packaging==24.0
pathspec==0.12.1
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
pluggy==1.4.0
psutil==5.9.8
ptyprocess==0.7.0
pycodestyle==2.11.1
pycparser==2.22
pydantic==1.10.15
pydyf==0.9.0
pyfiglet==1.0.2
Pygments==2.17.2
pyparsing==3.1.2
pypdf==4.1.0
pyphen==0.14.0
pytest==8.1.1
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
render50==9.2.7
reportlab==4.1.0
requests==2.31.0
setuptools==69.2.0
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.29
sqlparse==0.4.4
style50==2.10.3
submit50==3.1.4
termcolor==1.1.0
tinycss2==1.2.1
tqdm==4.66.2
typeguard==4.2.1
typing_extensions==4.10.0
tzlocal==5.2
urllib3==2.2.1
weasyprint==61.2
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.2
wheel==0.43.0
zopfli==0.2.3
